Expose `CasePathsMacroSupport` module
Gives downstream modules the ability to define `CasePathable` conformances.

Similar to how the `@Model` macro synthesizes `@Observable`, this provides a public way for
enum-based libraries to synthesize `@CasePathable`.
