+++ Package.swift
@@ -21,6 +21,10 @@ let package = Package(
       name: "CasePathsCore",
       targets: ["CasePathsCore"]
     ),
+    .library(
+      name: "CasePathsMacrosSupport",
+      targets: ["CasePathsMacrosSupport"]
+    ),
   ],
   dependencies: [
     .package(url: "https://github.com/swiftlang/swift-syntax", "509.0.0"..<"605.0.0"),
@@ -41,8 +45,8 @@ let package = Package(
         .product(name: "XCTestDynamicOverlay", package: "xctest-dynamic-overlay"),
       ]
     ),
-    .macro(
-      name: "CasePathsMacros",
+    .target(
+      name: "CasePathsMacrosSupport",
       dependencies: [
         .product(name: "SwiftSyntaxMacros", package: "swift-syntax"),
         .product(name: "SwiftCompilerPlugin", package: "swift-syntax"),
@@ -51,9 +55,21 @@ let package = Package(
         .product(name: "SwiftSyntaxBuilder", package: "swift-syntax"),
       ]
     ),
+    .macro(
+      name: "CasePathsMacros",
+      dependencies: [
+        "CasePathsMacrosSupport",
+        .product(name: "SwiftSyntaxMacros", package: "swift-syntax"),
+        .product(name: "SwiftCompilerPlugin", package: "swift-syntax"),
+        .product(name: "SwiftSyntax", package: "swift-syntax"),
+      ]
+    ),
     .testTarget(
       name: "CasePathsTests",
-      dependencies: ["CasePaths"]
+      dependencies: [
+        "CasePaths",
+        "CasePathsMacrosSupport",
+      ]
     ),
   ],
   swiftLanguageModes: [.v6]
@@ -75,6 +91,7 @@ if ProcessInfo.processInfo.environment["OMIT_MACRO_TESTS"] == nil {
       name: "CasePathsMacrosTests",
       dependencies: [
         "CasePathsMacros",
+        "CasePathsMacrosSupport",
         .product(
           name: "MacroTesting",
           package: "swift-macro-testing"
